+++ src/backend/src/models/costumers.ts
@@ -1,6 +1,7 @@
 "use strict";
 import { Model } from "sequelize";
 import { CostumersAttrebues } from "./Atterbuites/Costumers";
+import { IpTvType } from "./Atterbuites/Product";
 
 
 module.exports = (sequelize: any, DataTypes: any) => {
@@ -11,6 +12,7 @@ module.exports = (sequelize: any, DataTypes: any) => {
     Email!: string;
     referenceSite!: string;
     language!: string;
+    type!: IpTvType;
     bought!: boolean;
     bought_at!: Date;
     pendding!: boolean;
@@ -42,6 +44,11 @@ module.exports = (sequelize: any, DataTypes: any) => {
       bought: {
         type: DataTypes.BOOLEAN,
         defaultValue: false,
+        allowNull: false,
+      },
+      type: {
+        type: DataTypes.ENUM(...Object.values(IpTvType)),
+        defaultValue: IpTvType.Basic,
         allowNull: false,
       },
       bought_at: {