+++ changes.md
@@ -1,2 +1,13 @@
-refrence:
-- table titles
+## WORKING PROGRESS ##
+   - none
+
+## TESTING ###
+   - none
+
+## DONE ###
+   - pendding_costumer before Send email check for pendding type
+   - referenceTab on product
+   - pendding product dont show by defult
+   - penddng + reference flter on costumer returns [] 
+   - show costumer purchsed type on table
+   - add reference on main stats